Refactor type imports to optimize bundle size
