+++ utils/cost-calculations.ts
@@ -2,7 +2,7 @@
  * Enhanced cost calculation functions for date-based Betriebskosten system
  */
 
-import { Mieter } from "@/lib/data-fetching";
+import type { Mieter } from "@/lib/types";
 import { calculateTenantOccupancy, TenantOccupancy } from "./date-calculations";
 
 /**
@@ -15,7 +15,7 @@ export function calculateProFlächeDistribution(
   enddatum: string
 ): Record<string, { amount: number; occupancyDays: number; totalDays: number }> {
   const distribution: Record<string, { amount: number; occupancyDays: number; totalDays: number }> = {};
-  
+
   // Calculate total weighted area (area * occupancy ratio)
   let totalWeightedArea = 0;
   const tenantOccupancies: Array<{
@@ -28,22 +28,22 @@ export function calculateProFlächeDistribution(
     const occupancy = calculateTenantOccupancy(tenant, startdatum, enddatum);
     const apartmentArea = tenant.Wohnungen?.groesse || 0;
     const weightedArea = apartmentArea * occupancy.occupancyRatio;
-    
+
     totalWeightedArea += weightedArea;
     tenantOccupancies.push({ tenant, occupancy, weightedArea });
   });
-  
+
   // Distribute costs proportionally
   tenantOccupancies.forEach(({ tenant, occupancy, weightedArea }) => {
     const amount = totalWeightedArea > 0 ? (weightedArea / totalWeightedArea) * totalCost : 0;
-    
+
     distribution[tenant.id] = {
       amount,
       occupancyDays: occupancy.occupancyDays,
       totalDays: Math.ceil((new Date(enddatum).getTime() - new Date(startdatum).getTime()) / (1000 * 3600 * 24)) + 1
     };
   });
-  
+
   return distribution;
 }
 
@@ -57,7 +57,7 @@ export function calculateProMieterDistribution(
   enddatum: string
 ): Record<string, { amount: number; occupancyDays: number; totalDays: number }> {
   const distribution: Record<string, { amount: number; occupancyDays: number; totalDays: number }> = {};
-  
+
   // Calculate total occupancy days across all tenants
   let totalOccupancyDays = 0;
   const tenantOccupancies: Array<{
@@ -70,20 +70,20 @@ export function calculateProMieterDistribution(
     totalOccupancyDays += occupancy.occupancyDays;
     tenantOccupancies.push({ tenant, occupancy });
   });
-  
+
   // Distribute costs based on occupancy days
   const totalPeriodDays = Math.ceil((new Date(enddatum).getTime() - new Date(startdatum).getTime()) / (1000 * 3600 * 24)) + 1;
-  
+
   tenantOccupancies.forEach(({ tenant, occupancy }) => {
     const amount = totalOccupancyDays > 0 ? (occupancy.occupancyDays / totalOccupancyDays) * totalCost : 0;
-    
+
     distribution[tenant.id] = {
       amount,
       occupancyDays: occupancy.occupancyDays,
       totalDays: totalPeriodDays
     };
   });
-  
+
   return distribution;
 }
 
@@ -97,7 +97,7 @@ export function calculateProWohnungDistribution(
   enddatum: string
 ): Record<string, { amount: number; occupancyDays: number; totalDays: number }> {
   const distribution: Record<string, { amount: number; occupancyDays: number; totalDays: number }> = {};
-  
+
   // Group tenants by apartment and calculate total occupancy per apartment
   const apartmentOccupancy: Record<string, {
     tenants: Mieter[];
@@ -130,8 +130,8 @@ export function calculateProWohnungDistribution(
   const totalPeriodDays = Math.ceil((new Date(enddatum).getTime() - new Date(startdatum).getTime()) / (1000 * 3600 * 24)) + 1;
 
   Object.entries(apartmentOccupancy).forEach(([wohnungId, apt]) => {
-    const apartmentShare = totalWeightedOccupancy > 0 
-      ? (apt.totalOccupancyDays / totalWeightedOccupancy) * totalCost 
+    const apartmentShare = totalWeightedOccupancy > 0
+      ? (apt.totalOccupancyDays / totalWeightedOccupancy) * totalCost
       : 0;
 
     // Split apartment share equally among tenants in the apartment
@@ -139,7 +139,7 @@ export function calculateProWohnungDistribution(
 
     apt.tenants.forEach(tenant => {
       const occupancy = calculateTenantOccupancy(tenant, startdatum, enddatum);
-      
+
       distribution[tenant.id] = {
         amount: costPerTenant,
         occupancyDays: occupancy.occupancyDays,
@@ -147,7 +147,7 @@ export function calculateProWohnungDistribution(
       };
     });
   });
-  
+
   return distribution;
 }
 
@@ -166,17 +166,17 @@ export function calculateNachRechnungDistribution(
   tenants.forEach(tenant => {
     const occupancy = calculateTenantOccupancy(tenant, startdatum, enddatum);
     const individualAmount = individualAmounts[tenant.id] || 0;
-    
+
     // Apply the individual amount proportionally to occupancy
     const amount = individualAmount * occupancy.occupancyRatio;
-    
+
     distribution[tenant.id] = {
       amount,
       occupancyDays: occupancy.occupancyDays,
       totalDays: totalPeriodDays
     };
   });
-  
+
   return distribution;
 }
 
@@ -191,7 +191,7 @@ export function calculateWaterCostDistribution(
   enddatum: string
 ): Record<string, { amount: number; occupancyDays: number; totalDays: number; consumption?: number }> {
   const distribution: Record<string, { amount: number; occupancyDays: number; totalDays: number; consumption?: number }> = {};
-  
+
   // Calculate total weighted consumption (consumption * occupancy ratio)
   let totalWeightedConsumption = 0;
   const tenantData: Array<{
@@ -205,19 +205,19 @@ export function calculateWaterCostDistribution(
     const occupancy = calculateTenantOccupancy(tenant, startdatum, enddatum);
     const consumption = waterReadings[tenant.id] || 0;
     const weightedConsumption = consumption * occupancy.occupancyRatio;
-    
+
     totalWeightedConsumption += weightedConsumption;
     tenantData.push({ tenant, occupancy, consumption, weightedConsumption });
   });
-  
+
   // Distribute costs based on weighted consumption
   const totalPeriodDays = Math.ceil((new Date(enddatum).getTime() - new Date(startdatum).getTime()) / (1000 * 3600 * 24)) + 1;
-  
+
   tenantData.forEach(({ tenant, occupancy, consumption, weightedConsumption }) => {
-    const amount = totalWeightedConsumption > 0 
-      ? (weightedConsumption / totalWeightedConsumption) * totalWaterCost 
+    const amount = totalWeightedConsumption > 0
+      ? (weightedConsumption / totalWeightedConsumption) * totalWaterCost
       : 0;
-    
+
     distribution[tenant.id] = {
       amount,
       occupancyDays: occupancy.occupancyDays,
@@ -225,6 +225,6 @@ export function calculateWaterCostDistribution(
       consumption
     };
   });
-  
+
   return distribution;
 }
